Fix all 134 SIMD spec failures: v128 globals, store_lane, dot overflow
- module.zig: handle v128.const (0xFD 0x0C) in init expressions
- store.zig: widen Global.value from u64 to u128 for v128 globals
- instance.zig: evalInitExpr returns u128, v128.const in init exprs
- vm.zig: global_get/set use pushV128/popV128 for full u128
- vm.zig: skip RegIR/JIT for v128 results (not just params)
- vm.zig: fix i32x4_dot_i16x8_s wrapping arithmetic (INT16_MIN overflow)
- Spec: 56399/56399 (100.0%), 0 failures

+++ src/store.zig
@@ -140,7 +140,7 @@ pub const Table = struct {
 
 /// A Wasm global variable.
 pub const Global = struct {
-    value: u64,
+    value: u128,
     valtype: ValType,
     mutability: Mutability,
 };
@@ -495,7 +495,7 @@ test "Store — addGlobal and getGlobal" {
         .mutability = .mutable,
     });
     const g = try store.getGlobal(addr);
-    try testing.expectEqual(@as(u64, 42), g.value);
+    try testing.expectEqual(@as(u128, 42), g.value);
     try testing.expectEqual(Mutability.mutable, g.mutability);
 }
